Fix: Fix `HookMatcher.timeout` type to `float` (#357)
According to https://www.schemastore.org/claude-code-settings.json, the
`timeout` type is `number`, not `integer`

+++ src/claude_agent_sdk/types.py
@@ -367,7 +367,7 @@ class HookMatcher:
     hooks: list[HookCallback] = field(default_factory=list)
 
     # Timeout in seconds for all hooks in this matcher (default: 60)
-    timeout: int | None = None
+    timeout: float | None = None
 
 
 # MCP Server config
